refactor: put generic math lang AST in separate file
